+++ src/chat/client.cljs
@@ -67,8 +67,14 @@
 (defn- admin? []
   (= (.-value (query "#author")) "hkimura"))
 
-(defn- replace-content [element message]
-  (append-html element (str "<p>" message "</p>")))
+;; (map message (sort-by :msg/sent-at #(compare %2 %1) messages))
+
+(defn- replace-content [element messages]
+  (let [msg (sort-by :timestamp #(compare %1 %2) messages)]
+    ;; this again. do not forget.
+    (set! (.-textContent element) "")
+    (.insertAdjacentHTML element "afterbegin" (str "<p>" msg "</p"))))
+
 
 (defn- load-messages [n]
   (go (let [response (<! (http/get (str "/api/load/" n)))